backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
pyJWT==2.6.0
python-dateutil==2.8.2
mysqlclient==1.4.6